5.1. Refactoring. Remove unnecessary action and connect store into Main component
