+++ src/store/action.js
@@ -1,6 +1,5 @@
 const ActionType = {
   SELECT_GENRE: `SELECT_GENRE`,
-  SELECT_FILMS: `SELECT_FILMS`,
 };
 
 
@@ -8,11 +7,6 @@ const ActionCreator = {
   selectGenre: (genre) => ({
     type: ActionType.SELECT_GENRE,
     payload: genre,
-  }),
-
-  selectFilms: (films) => ({
-    type: ActionType.SELECT_FILMS,
-    payload: films,
   }),
 };
